+++ backend/modules/locale/engine/model.php
@@ -27,7 +27,7 @@ class BackendLocaleModel
 	public static function buildCache($language, $application)
 	{
 		// get db
-		$db = BackendModel::getDB();
+		$db = BackendModel::getContainer()->get('database');
 
 		// get types
 		$types = $db->getEnumValues('locale', 'type');
@@ -217,7 +217,7 @@ class BackendLocaleModel
 		$idPlaceHolders = array_fill(0, count($ids), '?');
 
 		// delete records
-		BackendModel::getDB(true)->delete('locale', 'id IN (' . implode(', ', $idPlaceHolders) . ')', $ids);
+		BackendModel::getContainer()->get('database')->delete('locale', 'id IN (' . implode(', ', $idPlaceHolders) . ')', $ids);
 
 		// rebuild cache
 		self::buildCache(BL::getWorkingLanguage(), 'backend');
@@ -232,7 +232,7 @@ class BackendLocaleModel
 	 */
 	public static function exists($id)
 	{
-		return (bool) BackendModel::getDB()->getVar(
+		return (bool) BackendModel::getContainer()->get('database')->getVar(
 			'SELECT 1
 			 FROM locale
 			 WHERE id = ?
@@ -262,7 +262,7 @@ class BackendLocaleModel
 		$id = ($id !== null) ? (int) $id : null;
 
 		// get db
-		$db = BackendModel::getDB();
+		$db = BackendModel::getContainer()->get('database');
 
 		// return
 		if($id !== null) return (bool) $db->getVar(
@@ -273,7 +273,7 @@ class BackendLocaleModel
 			array($name, $type, $module, $language, $application, $id)
 		);
 
-		return (bool) BackendModel::getDB()->getVar(
+		return (bool) BackendModel::getContainer()->get('database')->getVar(
 			'SELECT 1
 			 FROM locale
 			 WHERE name = ? AND type = ? AND module = ? AND language = ? AND application = ?
@@ -291,7 +291,7 @@ class BackendLocaleModel
 	public static function get($id)
 	{
 		// fetch record from db
-		$record = (array) BackendModel::getDB()->getRecord('SELECT * FROM locale WHERE id = ?', array((int) $id));
+		$record = (array) BackendModel::getContainer()->get('database')->getRecord('SELECT * FROM locale WHERE id = ?', array((int) $id));
 
 		// actions are urlencoded
 		if($record['type'] == 'act') $record['value'] = urldecode($record['value']);
@@ -317,7 +317,7 @@ class BackendLocaleModel
 		$language = (string) $language;
 		$application = (string) $application;
 
-		return BackendModel::getDB()->getVar(
+		return BackendModel::getContainer()->get('database')->getVar(
 			'SELECT l.id
 			 FROM locale AS l
 			 WHERE name = ? AND type = ? AND module = ? AND language = ? AND application = ?',
@@ -332,7 +332,7 @@ class BackendLocaleModel
 	 */
 	private static function getLabelsFromBackendNavigation()
 	{
-		return (array) BackendModel::getDB()->getColumn('SELECT label FROM backend_navigation');
+		return (array) BackendModel::getContainer()->get('database')->getColumn('SELECT label FROM backend_navigation');
 	}
 
 	/**
@@ -387,7 +387,7 @@ class BackendLocaleModel
 		foreach((array) $lbl as $label) $used['lbl'][$label] = array('files' => array('<small>used in navigation</small>'), 'module_specific' => array());
 
 		// get labels from table
-		$lbl = (array) BackendModel::getDB()->getColumn('SELECT label FROM modules_extras');
+		$lbl = (array) BackendModel::getContainer()->get('database')->getColumn('SELECT label FROM modules_extras');
 		foreach((array) $lbl as $label) $used['lbl'][$label] = array('files' => array('<small>used in database</small>'), 'module_specific' => array());
 
 		// loop files
@@ -930,7 +930,7 @@ class BackendLocaleModel
 		foreach($types as $key => $val) $types[$key] = '\'' . $val . '\'';
 
 		// get db
-		$db = BackendModel::getDB();
+		$db = BackendModel::getContainer()->get('database');
 
 		// build the query
 		$query =
@@ -1092,7 +1092,7 @@ class BackendLocaleModel
 	public static function getTypesForDropDown()
 	{
 		// fetch types
-		$types = BackendModel::getDB()->getEnumValues('locale', 'type');
+		$types = BackendModel::getContainer()->get('database')->getEnumValues('locale', 'type');
 
 		// init
 		$labels = $types;
@@ -1112,7 +1112,7 @@ class BackendLocaleModel
 	public static function getTypesForMultiCheckbox()
 	{
 		// fetch types
-		$aTypes = BackendModel::getDB()->getEnumValues('locale', 'type');
+		$aTypes = BackendModel::getContainer()->get('database')->getEnumValues('locale', 'type');
 
 		// init
 		$labels = $aTypes;
@@ -1165,7 +1165,7 @@ class BackendLocaleModel
 		if($date === null) $date = BackendModel::getUTCDate();
 
 		// get database instance
-		$db = BackendModel::getDB();
+		$db = BackendModel::getContainer()->get('database');
 
 		// possible values
 		$possibleApplications = array('frontend', 'backend');
@@ -1281,7 +1281,7 @@ class BackendLocaleModel
 		if($item['type'] == 'act' && urldecode($item['value']) != $item['value']) $item['value'] = SpoonFilter::urlise($item['value']);
 
 		// insert item
-		$item['id'] = (int) BackendModel::getDB(true)->insert('locale', $item);
+		$item['id'] = (int) BackendModel::getContainer()->get('database')->insert('locale', $item);
 
 		// rebuild the cache
 		self::buildCache($item['language'], $item['application']);
@@ -1301,7 +1301,7 @@ class BackendLocaleModel
 		if($item['type'] == 'act' && urldecode($item['value']) != $item['value']) $item['value'] = SpoonFilter::urlise($item['value']);
 
 		// update category
-		$updated = BackendModel::getDB(true)->update('locale', $item, 'id = ?', array($item['id']));
+		$updated = BackendModel::getContainer()->get('database')->update('locale', $item, 'id = ?', array($item['id']));
 
 		// rebuild the cache
 		self::buildCache($item['language'], $item['application']);