+++ backend/modules/pages/engine/model.php
@@ -67,7 +67,7 @@ class BackendPagesModel
 		$levels = self::getTree(array(0), null, 1, $language);
 
 		// get extras
-		$extras = (array) BackendModel::getDB()->getRecords(
+		$extras = (array) BackendModel::getContainer()->get('database')->getRecords(
 			'SELECT i.id, i.module, i.action
 			 FROM modules_extras AS i
 			 WHERE i.type = ? AND i.hidden = ?',
@@ -75,7 +75,7 @@ class BackendPagesModel
 		);
 
 		// get widgets
-		$widgets = (array) BackendModel::getDB()->getRecords(
+		$widgets = (array) BackendModel::getContainer()->get('database')->getRecords(
 			'SELECT i.id, i.module, i.action
 			 FROM modules_extras AS i
 			 WHERE i.type = ? AND i.hidden = ?',
@@ -345,7 +345,7 @@ class BackendPagesModel
 		$links = array();
 
 		// init var
-		$cachedTitles = (array) BackendModel::getDB()->getPairs(
+		$cachedTitles = (array) BackendModel::getContainer()->get('database')->getPairs(
 			'SELECT i.id, i.navigation_title
 			 FROM pages AS i
 			 WHERE i.id IN(' . implode(',', array_keys($keys)) . ')
@@ -469,7 +469,7 @@ class BackendPagesModel
 		$language = ($language === null) ? BackendLanguage::getWorkingLanguage() : (string) $language;
 
 		// get db
-		$db = BackendModel::getDB(true);
+		$db = BackendModel::getContainer()->get('database');
 
 		// get record
 		$page = self::get($id, $revisionId, $language);
@@ -523,7 +523,7 @@ class BackendPagesModel
 		$language = BackendLanguage::getWorkingLanguage();
 
 		// exists?
-		return (bool) BackendModel::getDB()->getVar(
+		return (bool) BackendModel::getContainer()->get('database')->getVar(
 			'SELECT 1
 			 FROM pages AS i
 			 WHERE i.id = ? AND i.language = ? AND i.status IN (?, ?)
@@ -551,7 +551,7 @@ class BackendPagesModel
 		$language = ($language === null) ? BackendLanguage::getWorkingLanguage() : (string) $language;
 
 		// get page (active version)
-		$return = (array) BackendModel::getDB()->getRecord(
+		$return = (array) BackendModel::getContainer()->get('database')->getRecord(
 			'SELECT i.*, UNIX_TIMESTAMP(i.publish_on) AS publish_on, UNIX_TIMESTAMP(i.created_on) AS created_on, UNIX_TIMESTAMP(i.edited_on) AS edited_on,
 			 IF(COUNT(e.id) > 0, "Y", "N") AS has_extra,
 			 GROUP_CONCAT(b.extra_id) AS extra_ids
@@ -607,7 +607,7 @@ class BackendPagesModel
 		$language = ($language === null) ? BackendLanguage::getWorkingLanguage() : (string) $language;
 
 		// get page (active version)
-		return (array) BackendModel::getDB()->getRecords(
+		return (array) BackendModel::getContainer()->get('database')->getRecords(
 			'SELECT b.*, UNIX_TIMESTAMP(b.created_on) AS created_on, UNIX_TIMESTAMP(b.edited_on) AS edited_on
 			 FROM pages_blocks AS b
 			 INNER JOIN pages AS i ON b.revision_id = i.revision_id
@@ -629,7 +629,7 @@ class BackendPagesModel
 		$tagId = (int) $tagId;
 
 		// get the items
-		$items = (array) BackendModel::getDB()->getRecords(
+		$items = (array) BackendModel::getContainer()->get('database')->getRecords(
 			'SELECT i.id AS url, i.title AS name, mt.module
 			 FROM modules_tags AS mt
 			 INNER JOIN tags AS t ON mt.tag_id = t.id
@@ -657,7 +657,7 @@ class BackendPagesModel
 		$pageId = (int) $pageId;
 
 		// get child
-		$childId = (int) BackendModel::getDB()->getVar(
+		$childId = (int) BackendModel::getContainer()->get('database')->getVar(
 			'SELECT i.id
 			 FROM pages AS i
 			 WHERE i.parent_id = ? AND i.status = ? AND i.language = ?
@@ -735,7 +735,7 @@ class BackendPagesModel
 		$id = (int) $id;
 		$language = ($language === null) ? BackendLanguage::getWorkingLanguage() : (string) $language;
 
-		return (int) BackendModel::getDB()->getVar(
+		return (int) BackendModel::getContainer()->get('database')->getVar(
 			'SELECT revision_id
 			 FROM pages AS i
 			 WHERE i.id = ? AND i.language = ? AND i.status != ?',
@@ -750,7 +750,7 @@ class BackendPagesModel
 	 */
 	public static function getMaximumBlockId()
 	{
-		return (int) BackendModel::getDB()->getVar(
+		return (int) BackendModel::getContainer()->get('database')->getVar(
 			'SELECT MAX(i.id) FROM pages_blocks AS i'
 		);
 	}
@@ -767,7 +767,7 @@ class BackendPagesModel
 		$language = ($language !== null) ? (string) $language : BackendLanguage::getWorkingLanguage();
 
 		// get the maximum id
-		$maximumMenuId = (int) BackendModel::getDB()->getVar(
+		$maximumMenuId = (int) BackendModel::getContainer()->get('database')->getVar(
 			'SELECT MAX(i.id) FROM pages AS i WHERE i.language = ?',
 			array($language)
 		);
@@ -793,7 +793,7 @@ class BackendPagesModel
 		$language = ($language !== null) ? (string) $language : BackendLanguage::getWorkingLanguage();
 
 		// get the maximum sequence inside a certain leaf
-		return (int) BackendModel::getDB()->getVar(
+		return (int) BackendModel::getContainer()->get('database')->getVar(
 			'SELECT MAX(i.sequence)
 			 FROM pages AS i
 			 WHERE i.language = ? AND i.parent_id = ?',
@@ -969,7 +969,7 @@ class BackendPagesModel
 		$language = ($language !== null) ? (string) $language : BackendLanguage::getWorkingLanguage();
 
 		// get data
-		$data[$level] = (array) BackendModel::getDB()->getRecords(
+		$data[$level] = (array) BackendModel::getContainer()->get('database')->getRecords(
 			'SELECT
 			 	i.id, i.title, i.parent_id, i.navigation_title, i.type, i.hidden, i.data,
 				m.url, m.data AS meta_data,
@@ -1164,7 +1164,7 @@ class BackendPagesModel
 		if($parentId == 0 || $parentId == 1 || $parentId == 2 || $parentId == 3 || $parentId == 4) $parentIds = array(0, 1, 2, 3, 4);
 
 		// get db
-		$db = BackendModel::getDB();
+		$db = BackendModel::getContainer()->get('database');
 
 		// no specific id
 		if($id === null)
@@ -1264,7 +1264,7 @@ class BackendPagesModel
 	 */
 	public static function insert(array $page)
 	{
-		return (int) BackendModel::getDB(true)->insert('pages', $page);
+		return (int) BackendModel::getContainer()->get('database')->insert('pages', $page);
 	}
 
 	/**
@@ -1275,7 +1275,7 @@ class BackendPagesModel
 	public static function insertBlocks(array $blocks)
 	{
 		// get db
-		$db = BackendModel::getDB(true);
+		$db = BackendModel::getContainer()->get('database');
 
 		// loop blocks
 		foreach($blocks as $block)
@@ -1304,7 +1304,7 @@ class BackendPagesModel
 		$language = ($language === null) ? BackendLanguage::getWorkingLanguage() : (string) $language;
 
 		// get db
-		$db = BackendModel::getDB(true);
+		$db = BackendModel::getContainer()->get('database');
 
 		// reset type of drop for special pages
 		if($droppedOn == 1) $typeOfDrop = 'inside';
@@ -1444,7 +1444,7 @@ class BackendPagesModel
 	public static function update(array $page)
 	{
 		// get db
-		$db = BackendModel::getDB(true);
+		$db = BackendModel::getContainer()->get('database');
 
 		// update old revisions
 		if($page['status'] != 'draft') $db->update('pages', array('status' => 'archive'), 'id = ? AND language = ?', array((int) $page['id'], $page['language']));
@@ -1507,7 +1507,7 @@ class BackendPagesModel
 		$newTemplate['data'] = @unserialize($newTemplate['data']);
 
 		// fetch all pages
-		$pages = (array) BackendModel::getDB()->getRecords(
+		$pages = (array) BackendModel::getContainer()->get('database')->getRecords(
 			'SELECT *
 			 FROM pages
 			 WHERE template_id = ? AND status IN (?, ?)',